added -jpa adding error handling

+++ src/main/java/com/cordero/fullstack/user/UserService.java
@@ -15,10 +15,17 @@ public class UserService {
     }
 
     public User createUser(User user) {
+        var u =userRepository.findUserByEmail(user.getEmail()).isPresent();
+        if (userRepository.findUserByEmail(user.getEmail()).isPresent()) {
+            throw new IllegalStateException(String.format("A user already exist with this email [%s]", user.getEmail()));
+        }
         return userRepository.save(user);
     }
 
     public void deleteUser(Long userId) {
+        if (userRepository.existsById(userId)) {
+            throw new IllegalStateException(String.format("User not found with Id [%s]", userId));
+        }
         userRepository.deleteById(userId);
     }
 }
